+++ src/shared/lib/index.ts
@@ -1,0 +1,5 @@
+export { email as isEmail } from './regs';
+export { password as isPassword } from './regs';
+export { phoneNumber as isPhoneNumber } from './regs';
+export { formatMinute } from './formatDate';
+export { formatMonth } from './formatDate';
